Moved cereal registration to src files
- Moved polymorphic cereal registration macros to source files to reduce
  code size.
- Moved to XML serialization due to issues with new version of Cereal
- Fixed CMake config for 64-bit
- Fixed 64-bit warnings (added explicit casts for size_t values)
- Applied singleton pattern to UnitFactory
- Added serialization unit test
- Increased oversampling of ladder filters to 8x

+++ VOSIMLib/include/MemoryPool.h
@@ -56,7 +56,7 @@ class MemoryPool
 	std::vector<Chunk*> m_freeChunks;
 	std::vector<Chunk*> m_usedChunks;
 	Chunk m_head;
-	int m_bytesUsed;
+	size_t m_bytesUsed;
 public:
 	MemoryPool(size_t a_numBytes);
